Updated /assets/css/{danhixon.scss,style.scss} and moved outdated /assets/css/default.css into /ignore/assets/css/
Changed import lines to be for jp-uberduper in /assets/css/danhixon.scss, and jp-minimal in /assets/css/style.scss (but added an extra, commented-out import line for jp-leapday, above the import for jp-minimal, because that's what it was referencing up until now anyway).

NOTE: From now on, all work will be going into the jp-minimal theme, unless I see anything that needs tweaking in the others, but the intention is to use jp-minimal for all pages from now on!